refactor: Extract BatchSnoozeResult to its own file
