+++ app/src/main/java/com/infomaniak/mail/data/api/ApiRepository.kt
@@ -54,6 +54,8 @@ import com.infomaniak.mail.data.models.mailbox.*
 import com.infomaniak.mail.data.models.message.Message
 import com.infomaniak.mail.data.models.signature.Signature
 import com.infomaniak.mail.data.models.signature.SignaturesResult
+import com.infomaniak.mail.data.models.snooze.BatchSnoozeCancelResponse
+import com.infomaniak.mail.data.models.snooze.BatchSnoozeUpdateResponse
 import com.infomaniak.mail.data.models.thread.ThreadResult
 import com.infomaniak.mail.ui.newMessage.AiViewModel.Shortcut
 import com.infomaniak.mail.utils.AccountUtils
