+++ app/src/main/java/com/infomaniak/mail/data/models/snooze/BatchSnoozeCancelResponse.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.data.models
+package com.infomaniak.mail.data.models.snooze
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
