+++ app/src/main/java/com/infomaniak/mail/data/models/snooze/BatchSnoozeResult.kt
@@ -17,11 +17,14 @@
  */
 package com.infomaniak.mail.data.models.snooze
 
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
+import androidx.annotation.StringRes
+import com.infomaniak.mail.data.cache.mailboxContent.ImpactedFolders
 
-@Serializable
-class BatchSnoozeUpdateResponse : BatchSnoozeResponse {
-    @SerialName("updated")
-    override val processedUuids: List<String> = emptyList()
+sealed interface BatchSnoozeResult {
+    data class Success(val impactedFolders: ImpactedFolders) : BatchSnoozeResult
+    sealed interface Error : BatchSnoozeResult {
+        data object NoneSucceeded : Error
+        data class ApiError(@StringRes val translatedError: Int) : Error
+        data object Unknown : Error
+    }
 }
